Use ClientRequest instead of resteasy3.
Previously we were using annotations to setup a REST client (via
resteasy-client) for PNC in Dependency Analysis. This worked well except
when we deploy on JBoss EAP 6.4. JBoss EAP 6.4 uses and provides
resteasy2 libraries, but we need resteasy3 libraries for resteasy-client
to work.

Instead, we have decided to use `ClientRequest` class provided by
resteasy2 to write the REST client. It's a little bit more verbose than
resteasy-client, but it does the job and avoids the version conflict we
have right now.

The `PNC` object will throw some exceptions, if the REST call fails or
the configuration in common cannot be parsed.

The `pom.xml` was adjusted so that we have no more dependency on
resteasy-client, and we use resteasy v2 instead.

I have also decided to rename my 'model classes' package to `model`,
instead of using `entity` to avoid confusion.