Initial commit for Swagger
The swagger ui can be accessed from:
http://localhost:8080/reports-rest/swagger-ui/

Due to an issue with Swagger-Resteasy library not being able to read
JAXRS annotations from interfaces, we decided to remove the interface
and add those JAXRS annotations to the implemented classes instead. As
such, `Artifacts.java` is switched from an interface to a concrete
class, and `ArtifactsImpl.java` is removed. Same thing for `Root.java`.

The `pom.xml` is modified to add the swagger dependency.

The `web.xml` is modified to add the servlet that will serve the REST
documentation JSON data at
http://localhost:8080/reports-rest/rest/api-docs

The `webapp` folder now includes another folder, `swagger-ui`, that will
consume the documentation JSON data and provides a nice webview of the
REST endpoints.

Note that we are using v2.0.24 for `swagger-ui`, since the latest one
from master can't properly display the top-level list of apis available
right now.