Add implementation for DependencyTreeGenerator
This commit also removes the `GAV` and the `ArtifactTree` classes found
in the backend module, since the exact same model classes are already
implemented in the `communication` aprox models.

Note that `SCMLocator` asks for pom path, version, revision, and scm
url. The version string was added since this is needed by the
`AproxConnector` interface.