Adding unit tests for PNC interface
The `DAConfig` class has `@Setter` annotation to the fields to make mocking the
`Configuration.getConfig()` method easier.

The `common` module has the added `beans.xml` to activate Weld/CDI.

The `communication` module has the added dependencies 'mockito' and
'commons-logging'. commons-logging was added due to this stacktrace when
running the Mockito test without the dependency:

```
java.lang.NoClassDefFoundError: org/apache/commons/logging/LogFactory
        at org.apache.http.impl.client.CloseableHttpClient.<init>(CloseableHttpClient.java:58)
        at org.apache.http.impl.client.AbstractHttpClient.<init>(AbstractHttpClient.java:287)
        at org.apache.http.impl.client.DefaultHttpClient.<init>(DefaultHttpClient.java:147)
        at org.jboss.resteasy.client.core.executors.ApacheHttpClient4Executor.<init>(ApacheHttpClient4Executor.java:113)
        at sun.reflect.NativeConstructorAccessorImpl.newInstance0(Native Method)
        at sun.reflect.NativeConstructorAccessorImpl.newInstance(NativeConstructorAccessorImpl.java:62)
        at sun.reflect.DelegatingConstructorAccessorImpl.newInstance(DelegatingConstructorAccessorImpl.java:45)
        at java.lang.reflect.Constructor.newInstance(Constructor.java:422)
        at java.lang.Class.newInstance(Class.java:442)
        at org.jboss.resteasy.client.ClientRequest.getDefaultExecutor(ClientRequest.java:89)
        at org.jboss.resteasy.client.ClientRequest.<init>(ClientRequest.java:99)
        at org.jboss.da.communication.pnc.impl.PNCConnectorImpl.getClient(PNCConnectorImpl.java:31)
        at org.jboss.da.communication.PNCConnectorTest.shouldAddAuthenticationTokenToHeaderForAuthenticatedEndpoint(PNCConnectorTest.java:70)
```

I'm not sure why this stacktrace happens on the Mockito test only.

The `PNCConnectorImpl` class was changed so that both the `Configuration` and
`PNCAuthentication` objects can be injected.

The `PNCConnectorImplTest` class uses Mockito to inject mock `Configuration` and
`PNCAuthentication` objects to the `PNCConnectorImpl` object instance.