Fix the unmarshalling error in the reports-rest '/lookup/gav' endpoint
The unmarshalling error occured because the `GAV` class has a default
contructor that requires arguments.

Jackson requires the Java POJOs to have a 'no arguments' constructor to be
able to match the JSON structure to a JAVA POJO [1]. Since the `GAV`
class previously had no 'no arguments' constructor, the matching fails.

By adding the `@NoArgsConstructor` to the `GAV` class, a 'no arguments'
constructor is created, which makes Jackson happy.

No tests are provided for this commit. However in the future integration
tests will be added to those REST endpoints to catch these kinds of
errors.

The commit was verified manually using the Swagger UI [2].

Fixes DA-80
[1]: https://github.com/FasterXML/jackson-annotations#using-constructors-or-factory-methods
[2]: http://localhost:8080/reports-rest/swagger-ui/#!/reports/lookupGav