Fix delete of GAV from list
There are 2 `delete` functions that were defined in `listings.sh`.

The first `delete` function defined implemented the `delete` of a GAV
from the black/whitelist.

The second `delete` function defined overwrote the first `delete`
function, which broke the functionality of the `delete` cli option.

This commit fixes it by renaming the first `delete` function to
`deleteGAVFromList`, and making the necessary adjustment in the main
script.