[DA-117] pom.xml has wrong dependencies on jackson
Currently, our pom.xml contains 2 dependencies related to jackson:
- org.codehaus.jackson-jackson-core-asl-1.9.9
- com.fasterxml.jackson.core-jackson-databind-2.4.4

We can see that the groupId for the 2 jackson dependencies are different, due
to the migration of Jackson from Codehaus to Github (the new version is
released under version v2.x)

Ideally we should be using the latest version, however JBoss EAP 6.4 currently
use the Codehaus Jackson, which we should also do so that the Jackson
annotations we use are detected.

Therefore we should remove the v2.x Jackson dependency.

This commit replaces both dependencies with:
- org.codehaus.jackson-jackson-mapper-asl-1.9.9

As a result of the refactorings, it seems this has broken the
`GAVIntegrationTest.java` test for conversion to JSON. The test actually
had the wrong expected JSON output. The expected output is adjusted to
make the test pass.

Thanks to @rnc (Nick Cross) for pointing out this issue!