[DA-118] /lookup/gav endpoint has duplicate fields
Reply to '/lookup/gav' is currently:

```json
[{"groupId":"org.goots",
  "artifactId":"maven-plugin-test",
  "version":"1.0",
  "bestMatchVersion":null,
  "availableVersions":[],
  "blacklisted":false,
  "whitelisted":false,
  "groupId":"org.goots",
  "artifactId":"maven-plugin-test",
  "version":"1.0"}]
```
for a requests for "org.goots-maven-plugin-test-1.0"

Notice the duplicate groupId, artifactId, and version fields. Jackson is
getting confused with the Jackson annotations, and getters and setter
used in the Reports.java class.

Thanks to Nick Cross (@rnc) for reporting this!

This commit attempts to fix it by:

1. Add the `@JsonIgnore` annotation to the 'gav' field so that the
   getters and setters generated for 'gav' to also be ignored by Jackson
2. Remove the `@JsonUnwrapped` annotation because this confuses Swagger
3. Let the current `@XmlTransient` annotation on the `gav` field tell
   Jackson and Swagger to ignore that field for serialization.