Replace usage of null with Optional
AproxConnector
Returning null when GA is not found in repository is unnecessary,
because when GA exists in a repository, it exists at least in one version
there. Therefore we can return empty list for GA which is not in the
repository.

VersionFinder
Changed return value of getBuiltVersionsFor and lookupBuiltVersions to
Optional of previously returned types.

ReportsGenerator
Changed return value of getReport to Optional of ArtifactReport.
I couldn't find better solution (without using isPresent) because
addDependencyReports throws checked CommunicationException. Strams
doesn't work well with checked exceptions.

/reports/gav
Optional from ReportsGenerator layers is mapped to OK response, or if
empty then error message is returned.

/reports/lookup/gavs
Optional from VersionFinder is mapped to full LookupReport, or if empty
then partial LookupReport is returned.

Tests
Tests that were testing that null was returned, tests if isPresent is
false on returned Optional, other tests just call get on the Optional
and proceed like previously.