Replace usage of null with Optional #2
VersionFinder
Changed return value of getBestMatchVersionFor to Optional of String.

VersionLookupResult
Changed type of bestMatchVersion to Optional of String.

ArtifactReport
Changed type of bestMatchVersion to Optional of String.

Tests
Tests that were testing that null was returned, tests if isPresent is
false on returned Optional, other tests just call get on the Optional
and proceed like previously.