[DA-114] Use Cartographer for dependency analysis
This commit adds support for analysis of a maven project using
Cartographer.

The '/reports/scm' REST endpoint is also implemented. Note that the
available version, and best match version for each of the dependencies
in the are not yet provided in the REST endpoint, but will be in the
future.

`SCMRequest` was deleted since it was very similar to `SCMLocator`

Error Handling
--------------
1. The `PomAnalyzer` object will throw a `PomAnalysisException` for any
   errors that occur in the analysis.

2. The SCM checkout process will throw an `ScmException` error.

3. AproxConnector's `getDependencyTreeOfRevision` method will throw
   both the `ScmException` and the `PomAnalysisException` if something
   wrong happens while getting the tree of dependencies from an SCM

4. `ReportsGenerator` will also throw `ScmException` and
   `PomAnalysisException` if something wrong happens while getting the
   tree of dependencies from `AproxConnector`

5. The method that handle the '/reports/scm' endpoint will catch any
   `ScmException`, or `PomAnalysisException` and show the error message
   in the JSON object returned.

Known Issues in that commit
---------------------------
SCM checking out is hardcoded to 'git' right now