[DA-145][DA-148] Fix for projects analysis with cli
Some projects' dependencies cannot be analyzed using the 'da-cli.sh'
script (for pom-bw and pom-report) because the `mvn dependency:list`
command fails.

This typically happens when we are missing an artifact which only gets
produced during the package phase.

The fix is to capture the failure and recommend the user to run `mvn
clean install` before running the 'da-cli.sh' script again.