[DA-171] Know if Aprox can analyze a GAV or not.
Previously, when we asked Aprox to analyze a GAV for its dependencies,
we received the same result whether the GAV was:

- not present in Aprox and therefore cannot be analyzed
- present in Aprox and having no compile dependencies.

This commit fixes it by first checking if Aprox has the requested GAV in
its public repository. If yes, then proceed to the actual analysis.

However if the GAV is not in the public repository, a
`FindGAVDependencyException` is thrown.

Right now the `DependencyTreeGeneratorImpl` just catches the new
exception and returns an `Optional.empty()`. This will be really fixed
in DA-170 in another commit.

The JUnit tests are also adjusted to test/adapt for the new behaviour.