[DA-210] scm report does not show dependency tree
Due to commit ae10c3a9, the full depency tree of an SCM URL was not
being shown.

This commit was done so that we could anaylyze the keycloak project,
which contains a pom.xml which cannot be properly parsed. But this
commit had a flawed logic. It should actually analyze all the pom.xmls,
not just the top-level pom. Reason is we have to analyze all the
pom.xmls to get the full dependency tree.

So instead, we'll try our best to parse all the pom.xmls, and if we
can't, we'll just skip it and warn the user about it.