[DA-202] Investigate why lookup test is non deterministic
This non-deterministic behaviour, where the test:

Test RestApiReportsRemoteTest::testGavLookupList

fails might be due to the fact that we are using a parallel stream to
concurrently add values to an ArrayList. However, an ArrayList is not
thread-safe by default and will probably suffer from a race condition,
since addition to the ArrayList is not synchronized.

So instead the ArrayList is replaced by a CopyOnWriteArrayList, which is
thread-safe.