[DA-192] Move creating of products to UI in Product Import
Instead of trying to create a product, we'll just assume that the
product already exists.

The REST call now expects to get the productId only, instead of the
product name.

The REST call for the ProjectImport process has also been changed so
that it expects a projectId instead of a project name.

We'll still try to create a product version if it hasn't been created
yet.