Changes to better deserialize JSON using the DTO
This change applies to:

- LookupReport
- NPMLookupReport
- MavenLookupResult
- NPMLookupResult

The changes are required to properly deserialize the JSON into the
objects for clients using the DTOs to deserialize.

While a few stuff were tried, it looks like Jackson is confused by the
combination of the 'Builder' pattern and '@JsonUnwrapped' annotation and
doesn't know what to do. To keep it simple, this commit removes the
`@Builder` annotation, and uses the 'NoArgsConstructor' to make sure
Jackson can deserialize properly.

Tests are written to make sure that the deserialization process works.